refactor: converted utils to ts
